Create useReadReplies hook
