+++ src/components/CommentCard/CommentCardReplyList.tsx
@@ -2,30 +2,18 @@
 
 import clsx from 'clsx';
 import { type ReactNode } from 'react';
-import { useInfiniteQuery } from '@tanstack/react-query';
 
 import { REPLIES_FETCH_COUNT } from '@/lib/constants';
-import { QueryKey } from '@/lib/enums';
 import { type CommentWithRelationsAndRelationCountsAndUserReaction } from '@/lib/types';
 
 import { CommentCardSkeleton } from '../CommentCardSkeleton';
 
 import CommentCard from './CommentCard';
 import useCommentCard from './useCommentCard';
+import { useReadReplies } from '@/lib/hooks';
 
 const CommentCardReplyList = (): ReactNode => {
   const { comment } = useCommentCard();
-
-  // TODO
-  const getReplies = async ({ pageParam }: { pageParam: number }) => {
-    const response = await fetch(
-      `/api/posts/${comment?.postId}/comments/${comment?.id}/replies?cursor=${pageParam}`
-    );
-
-    if (!response.ok) throw new Error('Network response was not ok');
-
-    return response.json();
-  };
 
   const {
     data,
@@ -34,12 +22,7 @@ const CommentCardReplyList = (): ReactNode => {
     isFetchingNextPage,
     status,
     fetchNextPage,
-  } = useInfiniteQuery({
-    queryFn: getReplies,
-    queryKey: [QueryKey.REPLIES, comment?.postId, comment?.id],
-    initialPageParam: 0,
-    getNextPageParam: (lastPage) => lastPage.data.nextCursor,
-  });
+  } = useReadReplies(comment);
 
   const handleClick = (): void => {
     fetchNextPage();
